Moved rootset helper into OpTk
