+++ hat/examples/experiments/src/main/java/experiments/DependencyTree.java
@@ -24,14 +24,13 @@
  */
 package experiments;
 
-import hat.optools.RootSet;
+import hat.optools.OpTk;
 
 import java.lang.reflect.Method;
 import jdk.incubator.code.Op;
 import jdk.incubator.code.dialect.core.CoreOp;
 import jdk.incubator.code.CodeReflection;
 import java.util.Optional;
-import java.util.Set;
 import java.util.stream.Stream;
 
 public class DependencyTree {
@@ -162,10 +161,7 @@ public class DependencyTree {
     public static void main(String[] args) {
         CoreOp.FuncOp f = getFuncOp("f");
         System.out.println(f.toText());
-        RootSet rootSet = new RootSet(f.body().entryBlock().ops().stream());
-
-       // Set<Op> roots = RootSet.getRootSet(f.body().entryBlock().ops().stream());
-        f.body().entryBlock().ops().stream().filter(rootSet.set::contains).forEach(op -> {
+        OpTk.unwrappedRootsWithoutVarFuncDeclarationsOrYields(f.body().entryBlock()).forEach(op -> {
             System.out.println(op.toText());
         });
